requirements.txt
Django==5.0.7
mysqlclient==2.2.4
python-dotenv==1.0.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
holidays==0.59
reportlab==4.0.7
openpyxl==3.1.2
